VIC-7013 Move vision schedule/mode reasoning entirely into Vision Component (#3939)
VisionScheduleMediator is the sole owner of the vision mode schedules.
VisionComponent controls which modes are actually enabled.
VisionModes will only run if both scheduled and enabled.
Which VisionModes to process are passed in to VisionSystem via the VisionSystemInput. This removes the need for all the current/next schedules/modes.

VisionModes are now default enabled (not necessarily scheduled). This is because VisionComponent has control over what modes actual run and we want most modes to be capable of running by just having a someone subscribe to a VisionModeSchedule. Most dev/unused modes are now disabled in vision_config.json for the same reason and must be explicitly enabled.

VIC-7256 Added/updated some debug displays. The vision mode schedule display now shows the schedule for all modes rather than just the enabled ones (default schedules were not considered enabled so they were not being shown). Switches VisionMode debug display in the main webots status window to show all processed modes for the current image.