CLAI-264 Attempt to refresh JWT token on PermissionDenied (#4001)
If the robot does not have the correct time, requests can fail due
to invalid tokens because the robot doesn't know it needs to
refresh its token. This change attempts to fix that issue by
refreshing the token when a service hits a PermissionDenied error.

To do this, this change adds the `util.ErrorListener` interface
which the voice and jdoc services now accept as an option. This
injects the token client's error handler into those services, so
it can then respond to an error.

When a PermissionDenied error happens in response to a request,
the token client will then attempt to refresh the JWT token,
retrying with exponential backoff. If a second PermissionDenied
error is encountered, it will cease all retrying; if other errors
occur (say, connection issues), it will keep trying to refresh.