VIC-9840 Alexa code face (#4050)
VIC-9840 calls for displaying the pairing code, URL, a nice
Alexa image, and some explanatory text on Vector's face during
the Alexa pairing process.
This is a partial implementation of that ticket that displays the
code and URL in a very ugly format. This should enable more people
to start using robots with Alexa. I will leave the ticket open.