VIC-11196 Fix Fetch bug when no face is found (#4094)
The first stage of the FetchCube behavior looks for a face to bring
the cube to by delegating to the FindFaceAndThen behavior. A corner
case in FindFaceAndThen would cause the behavior to hang if no face
was found when searching for a face from an observed body during the
LookForFaceInMicDirection phase. This change prevents that hang by
providing a valid state transition for this particular case.