SAIP-1926 Add ramp-up/ramp-down phase to Robot Test Fleet
This implements two different ramping strategies for the fleet:

1) The robots across every container will ramp up linearly based on
   the uniquely assigned robot and task IDs. The total ramp-up and
   -down delays for the entire fleet can be individually configured.

2) Ramping patterns can be generated by setting a robot setup /
   tear-down delay that is unique for every robot inside the cluster
   (using an external script). This allows for arbitrary ramping
   shapes.

This also refactors the robot simulation code and adds more extensive
unit testing for (distributed) control of the robots fleet and test
options processing.

The unique test IDs are now generated using a combination of the
robot index (unique inside a single task/container) and the task ID
(which is unique inside the cluster).