Updates to build scripts (go and load test container)
Add command line parsing to container build script that allows
a.o. to skip container image upload to AWS ECR.

The go build script for the Linux/x86 target now returns an exit
code when the build fails.

The build container script will abort when the build fails (so that
the container build and upload step is skipped).

A local development mode is introduced that allows building for
docker-compose.