VIC-9820 Integrate lttng userspace tracing
Linux Trace Toolkit (lttng) is a lightweight tracing system that
can be used to record events (using tracepoint) or logs (using
tracelog). This commit adds the ankitrace library with support for
tracing the duration of the main loops in vic-anim, vic-robot and
vic-engine as well as a trace event when a fault code is triggered.
These trace points are disabled in the shipping build but turned on for
release and user builds. This means if you need to trace a shipping build
you will need to build with -DUSE_LTTNG=ON.

This commit also includes the graph_durations.py script that will
process a trace record and graph a histogram of the main loop duration
timings. This script is expected to be used as an example and requires
babeltrace python bindings which is best used from Linux. The trace
records can also be analyzed with tracecompass which is a cross platform
graphical environment.