[VIC-11293] Added transition backpack lights from streaming to intent active (#4161)
* [VIC-11293] Added transition backpack lights from streaming to intent active

[VIC-11293] [Voice Commands] polish backpack lights transition for "Working on it" lights

When “hey vector” behavior detects an intent is pending, it plays a backpack light transition (simply the same streaming lights) from streaming to the active intent feedack lights.  This hides the gap in lights between the two points.

There is the chance that if no behavior activates the pending intent, the transition lights would remain on, in which case we have a safety timeout and error message.  Transition lights are cleared along with the pending intent.

* [VIC-11293] Active Intent Feedback Transition Lights

+ changed active intent transition lights from blue to flashing white
+ removed redundant “clear intent” function