CLAI-277 Add hotword parameter for timezone (#4191)
We will, soon, want to send the timezone from robot settings up to
Houndify for knowledge graph requests. This change adds the ability
to include that info in the hotword message sent from the anim
process to the cloud process; once that's hooked up on the robot
side, the information will automatically be passed up to Chipper.