VIC-11314 Crash in SetParent/TransitionToCheckForCubeInFront (#4203)
Several of the cube behaviors were storing ObservableObject* to cubes
from BlockWorld for use in identifying a target object for continued
interaction. This was causing crashes when the underlying objects
were destroyed as a result of normal operation elsewhere on the
BlockWorld interface. These changes refactor the pertinent behaviors
to store ObjectID's and obtain pointers directly from BlockWorld to
the corresponding objects when needed.