VIC-11362 vic-switchboard doesn't need opencv
When trying to run vic-switchboard as the `net` user instead of root,
it would fail at launch after loading and initializing an opencv
shared library.  This shared library was trying to access something
under `/proc` and did not have permission to do so.  However,
`vic-switchboard` doesn't actually need `opencv` for anything.  This
change removes that dependency.