VIC-11428 Setting for backpack button func (#4174)
By default, a single press of the backpack button triggers the "hey
vector" wakeword. When alexa is enabled, we want the provide the user
with an option to set whether the backpack button triggers "hey vector"
or the "alexa" wakeword.
When the user enables alexa, the button functionality should change to
Alexa, unless the user switches it back.