VIC-11362 switchboard: Use sudo for restricted ops
We are going to run switchboard as `net:anki` instead of `root:anki`
to reduce the damage that it can do if a hack or exploit is found.
However, switchboard still needs to do things like start/stop an OTA
update, reboot the robot, remove all wifi configs, and restart wifi.
For these operations, switchboard will rely on `sudo` to execute these
restricted operations.