VIC-10255 Allow Alexa shipping build to be downgraded
Normally, for security reasons, the OTA client does NOT allow a
downgrade from a later OS version to an earlier version for shipping
builds.  The version number on these Alexa builds is `2.0.0.X`.  Our
current production build is `1.0.1.1768` and thus we cannot use an OTA
to downgrade these robots in the field.  After we get passed
certification with Amazon we are likely to want to downgrade these
robots to `1.1.0.X` or maybe `1.2.0.X`.  After that, they will again
have a configuration that does NOT allow downgrading.