VIC-8849 Call `sudo /sbin/reboot` from vic-gateway (#4249)
The OS now defines an anki-specific sudoers policy under `/etc/sudoers.d`.  This policy allows `vic-gateway` to run `/sbin/reboot` as `root` without a password in order to perform a clean soft reboot.  The `/sbin/reboot` command (`systemctl reboot`) *must* be run as `root`, which is why the previous change to add the `CAP_SYS_BOOT` capability was insufficient.  Note that general `sudo` capabilities are _not_ permitted.  The `sudoers` policy restricts permissions to a minimal subset of required utilities.