VIC-11602 Flash briefing fix (#4256)
Ensures that data are not added to mp3Buffer if it is
full. The flow is now
(1) download data until the buffer is full or the source reader is blocked
(2) decode it all
(3) sleep by a fraction of the time that was decoded
(4) repeat.
Because it needs to download more now before starting to decode, I also
reduced the time it sleeps after decoding so that it will start downloading
again sooner, to avoid starving wwise. Still not an ideal approach. The
tradeoff is that it will now need to download up to the buffer size
before starting to decode. We probably don't need to download the original
buffer size of 250k, so I decreased the buffer to 65k.

This also fixes an issue where streams initially contain a ton of header
info that the decoding lib wants to skip. With a cbs sports example file,
it gobbles up 26k for the first batch of samples. However, we were giving
it a kMaxReadSize of 8192, so it would not find samples, and not decode
until flush. While 8192 is the guaranteed min size of a contiguous region
in the buffer, in practice it can be much larger. This adds a getter to
the ring buff to check that size, and then uses it when determining how
much to feed the decoding lib. Also adds some tests.

Also doesn't decode unless there's 8k in the buffer to avoid squeaks