VIC-11708 switchboard: Use sudo for diagnostics
With the changes for VIC-11362, `vic-switchboard` no longer runs as
root.  When it executes the `/anki/bin/diagnostics-logger` script as the
`net:anki` user, the script fails at various points to create
directories and gather information.  `/anki/bin/diagnostics-logger`
needs root privilege to gather all of the diagnostic information.  With
this change, `vic-switchboard` will use `sudo` to run
`/anki/bin/diagnostics-logger`.  `/anki/bin/diagnostics-logger` is also
modified to make sure that `net:anki` is the owner of the created
files.  This allows `vic-switchboard` to read them back and send them to
`mac-client`.