VIC-11481: Syscon quiet power management (#4127)
Syscon changes (actually merged to vicos in https://github.com/anki/vicos-oelinux/pull/483) that 

1) "disables" all the encoders whenever all the motors aren't moving and 
2) lowers the sampling rate of the prox sensor in order to save power and extend off charger operating time. 

When "disabled", the encoders are actually run at a slower rate so that motion can be detected. When motion is detected, encoders are re-enabled and status bits indicating whether the head or lift was moved are sent to the engine via RobotState.status since the motor has likely become uncalibrated. This is used to selectively trigger BehaviorReactToUncalibratedHeadAndLift and force a recalibration. (Calibration is suppressed during the FistBump behavior since lift manipulation is expected, but should trigger when the behavior completes.)

The new prox sensor behavior:
a) On charger or off, either tread moving, 32Hz (1 of 6 ticks)
b) Off charger, no tread moving, 16Hz (1 of 12 ticks)
c) On charger, no tread moving, ~1Hz (1 of 200 ticks)

Prox sensor changes are not handled in this PR but will be addressed in VIC-11727 and VIC-11730.