Temporarily disable Alexa wake word processing when "Alexa" plays
This change adds an API (I don't think one already exists?) to
temporarily disable a speech recognizer. When the TTS media
player's internal Alexa recognizer detects that the TTS audio
contains an "Alexa", it marks the timestamps in the current audio,
and attempts to detect when we reach that timestamp in the course
of playback.

After passing the start of the "Alexa", we temporarily disable the
real Alexa recognizer's processing, and after passing the end, we
re-enable it. In testing, this prevents the "Alexa" in "my name is
Alexa" from triggering another wake word.