VIC-11799 Save clock value on button press
We don't have a real hardware clock.  Our hardware clock just counts the
number of seconds since power up, but is not set to the number of
seconds since an epoch (like Jan 1, 1970).  If we had a real battery
backed hardware clock that kept ticking even when the robot was off, we
could boot up the robot and have a pretty good idea of what the real
wall clock time was without having to do NTP.  To work around this, we
try to fake a hardware clock, by saving out the wall clock time to a
file every 15 minutes and during a full and proper systemd shutdown.
However, we also allow the user to bypass the proper shutdown and just
hold the button for a few seconds.  When that happens, we don't do that
final save out of the clock value to disk.  This change will cause us to
save the clock value to disk when the button is pressed and try to do it
a few more times as the user holds the button for a shutdown.  When the
robot boots back up, it will have a more accurate, but not perfect, idea
of what the current wall clock time should be.  Hopefully, the robot
will eventually be able to contact an NTP server and get a better value.

To note, after the user presses and holds the button and the robot
powers off, our hardware clock will stop ticking and reset to 0 as it no
longer has power to maintain its count.  Until the user powers on the
robot again, our saved wall clock value will fall behind the real wall
clock time.  If the user waits 1 hour to power on the robot, then when
the robot first boots up, he will falsely think the current time is 1
hour in the past.  Once he gets a response from an NTP server, he will
correct himself.  Applications that have to be certain about the wall
clock time should wait until OSState::IsWallTimeSynced() returns true.

See also:
https://ankiinc.atlassian.net/wiki/spaces/VD/pages/457474385/Victor+System+Clock

(cherry picked from commit 89ff8e93d42d75993a8c393e1b05fb9e0e51007a)
Signed-off-by: Stuart Eichert <seichert@anki.com>

Conflicts:
	robot/supervisor/src/cozmoBot.cpp