VIC-9664 VIC-9665 VIC-9666 VIC-11090 Onboarding 1p2 (#4321)
These changes clear out old (on-rails) onboarding code and define new
App <-> Robot messages and behaviors for an App-driven onboarding
paradigm which aims to drastically reduce the amount of meaningful
state synchronization required between the robot and app during
onboarding. Onboarding "phases" can now be added as modular, atomic
behavior entities the order and transitions of and between which are
driven entirely by the SetPhase command, and which run independently
to completion providing only a Complete and CompletionPercentage (for
behaviors which require progressive user interaction) state required
by the app. If the app disconnects the robot goes to the LookAtPhone
state and awaits reconnection. Timeouts are added to each phase and
globally to prevent users getting stuck in onboarding. This may
require tuning and polish to function entirely as intended.

Still TODO:
* App backward compatibility with Hail Mary style OB
* Come Here and Meet Victor Phases