VIC-11805 Play signed out mp3 if previously authed (#4363)
If the user was _ever_ authenticated, even on another boot, this
will enable the wakeword and play an mp3 when they use it.
A resulting change is that the alexa folder isn't deleted, but only
the files in it. I also confirmed signout works VIC-11850 with the new
system, but I suspect it worked with the old system as well.