VIC-11787 Rewrite ReadFileAsBinary without fstream
We don't need a stream when we are reading binary data from a file.  The
`ifstream` code is hard to read and understand what it will do in
failure cases.  This rewrite uses the older methods that work with file
descriptors to open, seek, and read bytes from a file.  In addition, the
signature of `ReadFileAsBinary` is expanded to allow the caller to
specify an offset and length for when only a portion of the file is
desired.