VIC-11787 Make switchboard key migration more defensive
To our knowledge, there have not been any failures in the field for
`vic-switchboard` when it migrates keys.  However, the previous code did
not do a lot of error checking and that made us feel shaky.  This
rewrite checks that each phase of the migration succeeds and logs an
error if it does not.