VIC-12010 developer logging is not the same as crash logging
Restores full logging to a separate SaveToFile logger, this is full parity
and unfiltered, and written on a background thread. Logs are written to
/tmp/logs/<name of process> and are capped to a maximum size before rolling
over. To avoid filling storage all logs are deleted on process start.

The logging information uses the same logging source as the victor logger,
aka the original dev logging by using a multi-logger.

Crash logging is only enabled for debug and release builds and is inhibited
on shipping builds by setting ANKI_NO_CRASHLOGGING=1.

Performance was checked by reporting the CozmoAPI.CozmoInstanceRunner.overtime
statistic for every frame and comparing with crash logging enabled and disabled,
no impact was reported aside from the usual drop in performance when logging
was directed to the developer logging.