VIC-12005 vic-cloud : Delete corrupt token file
The following bad loop was happening:
1. `vic-cloud` initialized and failed to parse a corrupted token file.
2. `vic-cloud` set fault code 851 and exited
3. `vic-cloud` was restarted by the fault-code-handler
4. Go to step 1

With this change, if `vic-cloud` will delete the token file if it fails
to parse it.  As a guard against having a partially written corrupt
token file, this change also first writes the token file to a temporary
file and then renames the temporary file to the token file.