VIC-11716 react to cliff maneuvers for better edge handling (#4375)
The robot now executes an arc turn action while playing the appropriate animation instead of executing the drive-track of the animation. This motion is safer to execute when on the cliff edge.

Additionally, fixed the sequence of actions to be:

Animation (Huh) -> Animation (Reaction1) -> Turn to Cliff (if needed) -> Animation (Reaction Backup) -> Turn to Cliff (mandatory)

the first turn is now optional, so that the animations can sequence better in some scenarios.