VIC-11771: Robot getting stuck in ReactToUncalibratedHeadAndLift (#4426)
CalibrateMotorAction was timing out because sometimes syscon reports blips in head encoder position that makes us think it's moving (for ~100ms every 200ms) even when it's not. It's not even being powered. This caused it to get stuck in the HCS_LOWER_HEAD state when a calibration was commanded. I removed the check for IsMoving() in that state since that's only there for the HCS_RAISE_HEAD state which only gets used in factory, so I removed that state as well.