VIC-11276 gaze averaging in world coordinates (#4391)
This is the first in a series of pull requests for gaze direction. This pull request contains changes for how the underlying gaze point on the ground plane is computed. This algorithm is very similar to the one that was used in eye contact. All of this functionality is behind a feature gate, which is currently turned off.

Specifically the new items are 

 - Correctly setting the head pose's rotation matrix
 - Piping through TrackedFace via FaceWorld (where the gaze direction instances live)
 - The algorithm in gaze direction (this is the meat of this PR).
 - The visualizations of the gaze points.