[ALEXA ONLY] VIC-11866 VIC-11845 Detect notch for alexa loop (#4437)
The playback recognizer is not recognizing "alarm set for" or
several other utterances that trigger the alexa recognizer.
This adds a band stop filter to TTS output, and then looks for the
fingerprint of that filter in the speaker input. If it is detected
when someone speaks the wakeword, the callback is ignored.
I tested this with the playback recognizer disabled. If alexa is at
her lowest volume, we sometimes still need the playback recognizer
because the input energy is so low.
There's still more tuning that can be done, especially at low volume.
Also adds fftw to the alexa branch.