VIC-11999 Onboarding ComeHere and MeetVictor Tweaks (#4445)
* Add intent whitelist to reactToVoiceCommand to allow "i dunno"
  reaction to unhandled intents

* Set simulated streaming duration as a parameter of TriggerResponse

* Redirect webots fake button press to anim process

* Fix streaming light start time bug
The changes to micDataSystem cause the streaming backpack lights to
eluminate with the earcon instead of when streaming starts. We were
getting away with this since behaviorReactToVoiceCommand was turning
on the lights from the engine process and WakeWordlessStreaming was
implemented independently, but simulation showed the delay. The json
file changes stop reactToVoiceCommand from commanding backpack lights
so the bug is visible, and subsequently visibly fixed by the changes.