VIC-11866 VIC-11845 Detect notch for alexa loop
The playback recognizer is not recognizing "alarm set for" or
several other utterances that trigger the alexa recognizer.
This adds a band stop filter to TTS output, and then looks for the
fingerprint of that filter in the speaker input. If it is detected
when someone speaks the wakeword, the callback is ignored.
I tested this with the playback recognizer disabled. If alexa is at
her lowest volume, we sometimes still need the playback recognizer
because the input energy is so low.
There's still more tuning that can be done, especially at low volume.
Uses kissfft (BSD), as header only instead of a lib for now since
we are considering moving to a better library.

(cherry picked from commit 3b3eb4758e35429d90458515a900a981f4e1e3ab)
Signed-off-by: Udo Richter <udo@anki.com>