VIC-11962 Fix Final Check Pull Cube Tab Feature
The onboarding behavior used to subscribe to cube connections at start
up. This meant that a cube connection could be held open in some cases
draining the battery, and, more critically, any cubes not de-tabbed
prior to robot boot would never be detected.

These changes remove the subscription from onboarding behaviors and
add logic to the cubeCommsComp to start a scan if it is asked for a
list of available cubes and has none. This should be separated into
discreet 'scan' and 'getAvailableCube' messages when able.