VIC-11824 Stop playback on invalid audio format (#4465)
Sometimes we receive URLs or AttachmentReaders that aren't in mp3
format. If the URLs are playlists, an SDK component seems to be able
to convert a playlist into a continuous attachment reader stream. But
again, if that resulting stream is not in mp3 format, there's no
decoder for other formats.
This change counts the number of bytes the mp3 decoder library skips
without a successful mp3 frame decoded. If it exceeds some limit,
playback is aborted and we call onPlaybackError().
I had hoped this would result in amazon sending us some response to
the user, but that's not the case. For now, we just exit alexa face.