VIC-11694 VIC-11844 VIC-12007 Mic mute updates (#4483)
VIC-11694 Muting/unmuting quickly doesn't play all animations. The
fix restarts an animation each time mute state changes.

VIC-11844 Persists mic mute across reboots. Touches or deletes a
persistent file "micMuted"

VIC-12007 Single press should start a wake word response but show
the unmute animation. This suppresses any get-in animation (for alexa
or hey vector) played in the anim process, and simultaneously starts
the unmute animation through the face info screen. Since AlexaComponent
and BehaviorAlexa, and UserIntentComponent and
BehaviorReactToVoiceCommand, are waiting for the get-ins to complete,
this also changes the callback structure to actually check if the
get-in started.
There are complications: This is the first time that a face screen
allows a wakeword. Also, the engine response to the trigger word is
determined by what's in the stack, and in this case, engine switches
into and out of BehaviorWait at the same time that anim is trying to
figure out how to handle wake word triggers. Engine and anim end up
fighting over whether there is a response to the wake word based on
the timing of messages transferred between them. To handle this in
the current architecture, messages from anim to engine DebugScreenMode
and TriggerWordDetected have new params to explicitly handle these edge
cases.
This sadly adds more tangles to two systems that are split between
processes, wake word responses split between anim and engine, and
face info screen's use of Wait in the BehaviorComponentMessageHandler.
Since we keep talking about fixing these systems, I tied logic to an
edge case param instead of generalizing.