VIC-12228 update-engine.py: Refactor downgrade protection
Remove restrictions on the build suffix, but check that the new
version is greater than the current version (should be 0.9.X for the
factory).  Also check that the new version is greater than the minimum
allowed.  The current minimum is 1.0.0.1741.  However, this will be
bumped up later, once we have a version of VicOS that supports the
Whiskey hardware.