VIC-11613 VIC-11541 VIC-12100 Feedback when signing in/out of alexa (#4485)
We used to steal the sign in/out intents out from underneath
BehaviorReactToVoiceCommand. For sign in, we'd then switch immediately
to a face info screen. For sign out, nothing would happen
behavior-wise, until BehaviorReactToVoiceCommand played "I can't do
that," since its intent was stolen.
This adds logic for pending/claimed sign in/out requests in
AlexaComponent, and a behavior to activate when there's one pending. If
no behavior is there to handle it, AlexaComponent completes the request
automatically.
When signing in, the behavior looks at a face if there's one in front,
or looks up otherwise, then the code face kicks in.
When signing out, it looks at a face or up, then plays an animation.
When using a voice command to sign in/out, the standard earcon is
played, but when using the app, it's silent. We should probably get
an animation with audio, at least for the signout case.

(cherry picked from commit 2bf667bc9456c92f0f98316b04761948d6c6099f)
Signed-off-by: David Woods <dwoods@anki.com>