VIC-11964 Mute shouldn't interrupt place on charger during boot
This change disable the "double press to Mute" until after the engine
has finished loading. Since the behavior stack gets reset upon exiting
the info face, all kinds of havok is wreaked if we allow this to
happen before the engine is fully initialized, so lets not (for now).

Note that there is still a corner case where entering the customer
care screen before the engine has loaded will break things upon
exiting the info face, but disabling the customer care screen for any
period of robot operation is not an acceptable solution. The good
news is that its a hard case to hit: the user must hold Vector on his
charger and the charger in the air during for this to happen, unless
something is going wrong during IMU calibration in which case the
customer care screen will likely be needed in any event.