VIC-12051 Onboarding should resume after Mute feedback
Since Mute and the customer care face are implemented in the anim
process, they must reconfigure the behavior stack in order to have
full/reliable control of the robot. This causes problems in onboarding
since the stack reset after the Mute animation would put the robot
into an onboarding phase which the App did not command, causing de-
synchronization that is problematic in the "App Authoritative"
paradigm. These changes make onboarding recover to its prior state if
interruption is detected (ie if the onboarding behavior exits due to
anything other than an App command or a timeout), so Mute and the
customer care face work normally.

A corner case remains: if the user puts the robot into the Mute/CC
screen and then presses a button in the App which sends a command to
the robot, the command will be missed, potentially causing desync.
The onboarding timeouts will still prevent getting stuck indefinitely
but this case should be appropriately handled via proper message
handling on the App and the robot. This is documented in VIC-12357.

(cherry picked from commit d8cb23afc45461fb718df9d5c47833ac03ba2c79)
Signed-off-by: David Woods <dwoods@anki.com>