VIC-12329: Sometimes goes to AskForHelp on an easily recoverable cliff detection. (#4566)
I once saw the robot on clear surface with both front sensor hanging over the edge. All it had to do was backup. The presence of TooManyRobotStops in the log indicates that numStops is incrementing more than it should. It should really only increment when ReactToCliff is already active and not potentially increase to a large number while the behavior is out of scope.

(cherry picked from commit a93bdbe16cbb32d303dbf8d7ebd94510328553ef)
Signed-off-by: David Woods <dwoods@anki.com>