VIC-12369: Sometimes backs up for a while following ReactToOnBack (#4599)
CoordinateWhileInAir will lock tracks anytime the robot is in the inAir state which was cutting off the BodyMotionKeyFrames that were being sent during the onBack behavior. This bug was likely introduced when we made a change to allow onBack and onFace reactions to continue even after no longer being the onBack and onFace OffTread states, respectively, so as to finish playing out the animation.