VIC-12438 Always handle incoming onboarding messages from App (#4606)
These changes add a component to handle onboarding related messages
properly even if the onboarding behavior is inactive. This comes up
with Mute and the CC screen in particular since the onboarding be-
havior gets kicked off the stack, reliable RPC's from the App were
going unanswered during those times, causing the App to hang.

These problems were also pertinent in the 1p0 backward compatibility
case, so that had to be restructured as a special case onboarding
phase within the new system.