VIC-11967 VIC-11948 Fix for alexa blank face (#4612)
I was using the var animsComplete = !IsControlDelegated() to check
if anims were still running, but then possibly delegating and
checking animsComplete again.
The fix also checks IsControlDelegated() to know when to cancel the
behavior. This should help timing-based bugs where the face disappears
for a few seconds when an alexa interaction ends, until the
next animation triggers.
There may be more instances of this bug.