VIC-11427 Fix media player leak if URL content was played (#4629)
If a `UrlContentToAttachmentConverter` is created, it's given a
shared pointer to the current media player and a shared pointer to
it is created, making a cycle. Without making sure the converter
is shut down and destroyed when the media player is shut down, that
cycle could prevent destruction of the media player.

This is not a _full_ fix for the bug - there's still an object that
isn't being freed, but this at least will help.

(cherry picked from commit c4cbbe5f4716f46fe8aa81b12ce1ff88a8a64f27)
Signed-off-by: David Woods <dwoods@anki.com>