VIC-12633 Reduce calls to FillBorder (#4668)
I saw a lot of SIGILLs in backtrace in which some thread's callstack
contained a call to FillBorder. This runs a one-step flood fill
algorithm. It's called multiple times per tick during exploring.
This reduces the frequency it is called. The robot might to
react to explored obstacles multiple times as a tradeoff.