VIC-12106 Dispatcher Behaviors Will Send Multiple ActiveFeature’s to App
When a dispatcher behavior’s delegate finishes, there is a tick in where there is no delegated behavior, so the dispatcher will return it’s own ActiveFeature if set.  This causes the app to display an unecessary feature string for a short time after a dispatched behavior finishes.  This was the case for all BasicVoiceCommand behaviors.

Solution:
+ dispatchers ONLY return an ActiveFeature when they have delegated something, and that value is value from the previously dispatched behavior