VIC-12801: Replace fake-hwclock-tick fork/exec with socket activation
This commit removes the Fork/Exec used to start the fake-hwclock-tick
service when the backpack button is pressed and replaces it with a
systemd socket activation FIFO at /run/fake-hwclock-cmd.

We found using fork/exec after mlock causes some issues around delays
and memory usage. The additional threads can add enough mlocked pages to
the process to bump into the memlock ulimit. To complicate matters the
clone call can take over 2ms to complete causing spine CRC errors. When
moving the fork to separate thread we also had to set the affinity so it
does not run on the same RT core. This all added enough complications to
ditch the entire approach.

The VERSION file is being updated because this change requires version
1.3.1 of vicos which adds socket activation support.