VIC-12624: Neural nets in engine process (#4671)
This is step one of removing vic-neuralnets process (and is the main step). 
With this PR, we go back to the original approach that the std::future inside NeuralNetRunner creates a model that runs (in a thread) in the same process instead of talking through a "messenger model" via file I/O to the vic-neuralnets process. 

This PR still leaves the messenger model in place (and in fact updates it to be part of cti_neuralnets and fit the same interface), because it will still be useful for "offboard" vision processing (e.g. the cloud or on a laptop). Thus it is renamed to an NeuralNets::OffboardModel.

As part of this rearranging, I also did a bunch of cleanup (like converting to LOG_X macros, from PRINT_NAMED_X).

Actually removing the build components for vic-neuralnets will be a separate, forthcoming PR.