picovoice: Add SpeechRecognizer toggle to console system
This changeset adds primitize support for changing speech recognizers at runtime.
To avoid delays that might happen during the init/loading phase of each recognizer,
we init all of the models during startup, but only update the "active" recognizer
based on the value of a console variable (defined in anim/SpeechRecognizer.System).

This also adds a wakeword selector and sensitivity param for PicoVoice for eval purposes.