picovoice: Add build support for mac + vicos platforms
Add mac libs, keyword files & required platform defines in order to build and use picovoice
on mac. For simplicity, we use a static lib on mac. If not for the hard-float workaround
we would also prefer to use a static lib for vicos.

This commit also add `unused` annotations to fix mac & vicos-shipping builds.