Error 15 ACTION_FAILED
During dock with charger, managed to hit the old bug where the dock action thinks it is already close enough to a preaction pose but a different driving action does not think it is close enough the the preaction pose goal.
Fixed by exposing DriveToObjectAction’s`DoPositionCheckOnPathCompletion` bool to DriveToAndMountChargerAction.