BI-1719 Import okaoVision 2016.03.23 from coretech-external
Previously, we got okaoVision from coretech-external.  This is not
something we expect to modify or build again, so it is easier to just
import it here directly.  To conserve space in our git repo, I only
imported the includes and libs that we need to build.  I left out the
documentation and libraries for Android architectures that we don't
use.  I also left out the iOS libraries.  If you need those, see
coretech-external.

squash! BI-1719 Import okaoVision 2016.03.23 from coretech-external