The Box: Person Detection (#4823)
Overhaul CountPeople behavior 
Flow: Starts waiting for motion, ticking in BehaviorUpdate until motion is observed. Then transitions to running a frame of the Person Classifier. If nothing found, goes back to waiting for motion. If a person is present, send an image to the offboard person detector in the cloud. If that finds something, display it and wait. Either way, go back to looking for motion.

* New state machine with vision modes controlled directly by behavior instead of VisionModeDependencies (so VisionModeDependencies still exists, but has not dependencies registered now)
* Note that behavior now subscribes to the old RobotObservedMotion events to decide when to run Person Classifier
* Drawing done by behavior (using new cached mirror mode images in AIWhiteboard)