BI-1719 Remove dependency on coretech-external
coretech-external contained a variety of dependencies that we no longer
used and was difficult to maintain.  I created independent build
projects for the pieces that we did use and added them explicitly to our
DEPS.  This should make it easier to add new external build artifacts in
the future without having to worry about keeping all of
coretech-external building.